refactor: 💡 move keyboard composables to feature folder
