+++ app/src/main/java/com/rickyhu/hushkeyboard/keyboard/ui/buttons/NotationKeyButton.kt
@@ -1,4 +1,4 @@
-package com.rickyhu.hushkeyboard.ui.keyboard.buttons
+package com.rickyhu.hushkeyboard.keyboard.ui.buttons
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
